Apply translation for Scrobble Menu and Select Tracks Menu. Use `{{TEXT(…)}}` in JavaScript files. But it doesn't based on XMLConverter. Just simple string replacement.
