+++ WebServer.py
@@ -25,6 +25,7 @@ except ImportError:
     import xml.etree.ElementTree as etree
 
 import Settings, ATVSettings
+from Settings import getTranslation
 from Debug import *  # dprint()
 import XMLConverter  # XML_PMS2aTV, XML_PlayVideo
 
@@ -73,6 +74,15 @@ class MyHandler(BaseHTTPRequestHandler):
                     options[parts[0]] = ''
                 else:
                     options[parts[0]] = urllib.unquote(parts[1])
+
+            accept_language = self.headers.get('Accept-Language', 'en')
+            accept_languages = re.findall('(\w{2}(?:[-_]\w{2})?)(?:;q=(\d+(?:\.\d+)?))?', accept_language)
+            accept_languages = [(lang.replace('-', '_'), float(quotient) if quotient else 1.) for (lang, quotient) in accept_languages]
+            accept_languages = sorted(accept_languages, key=itemgetter(1), reverse=True)
+            for language, quotient in accept_languages:
+                if os.path.exists(os.path.join(sys.path[0], 'assets', 'locales', language, 'plexconnect.mo')):
+                    options['PlexConnectATVLanguage'] = language
+                    break
                     
             dprint(__name__, 2, "cleaned path:\n{0}", self.path)
             dprint(__name__, 2, "request options:\n{0}", options)
@@ -94,7 +104,7 @@ class MyHandler(BaseHTTPRequestHandler):
                     dprint(__name__, 1, "serving application.js")
                     f = open(sys.path[0] + sep + "assets" + sep + "js" + sep + "application.js")
                     self.send_response(200)
-                    self.send_header('Content-type', 'text/html')
+                    self.send_header('Content-type', 'text/javascript')
                     self.end_headers()
                     self.wfile.write(f.read())
                     f.close()
@@ -105,9 +115,15 @@ class MyHandler(BaseHTTPRequestHandler):
                     dprint(__name__, 1, "serving  " + sys.path[0] + sep + "assets" + self.path.replace('/',sep))
                     f = open(sys.path[0] + sep + "assets" + self.path.replace('/',sep))
                     self.send_response(200)
-                    self.send_header('Content-type', 'text/html')
+                    self.send_header('Content-type', 'text/javascript')
                     self.end_headers()
-                    self.wfile.write(f.read())
+                    text_content = f.read()
+                    translation = getTranslation(options.get('PlexConnectATVLanguage', 'us'))
+                    for msgid in set(re.findall(r'\{\{TEXT\((.+?)\)\}\}', text_content)):
+                        # get translated text and encode to numeric entities
+                        msgstr = ''.join(r'\u{0:04x}'.format(ord(x)) for x in translation.ugettext(msgid))
+                        text_content = text_content.replace('{{TEXT(%s)}}' % msgid, msgstr)
+                    self.wfile.write(text_content)
                     f.close()
                     return
                 
@@ -136,14 +152,6 @@ class MyHandler(BaseHTTPRequestHandler):
                 # get everything else from XMLConverter - formerly limited to trailing "/" and &PlexConnect Cmds
                 if True:
                     dprint(__name__, 1, "serving .xml: "+self.path)
-                    accept_language = self.headers.get('Accept-Language', 'en')
-                    accept_languages = re.findall('(\w{2}(?:[-_]\w{2})?)(?:;q=(\d+(?:\.\d+)?))?', accept_language)
-                    accept_languages = [(lang.replace('-', '_'), float(quotient) if quotient else 1.) for (lang, quotient) in accept_languages]
-                    accept_languages = sorted(accept_languages, key=itemgetter(1), reverse=True)
-                    for language, quotient in accept_languages:
-                        if os.path.exists(os.path.join(sys.path[0], 'assets', 'locales', language, 'plexconnect.mo')):
-                            options['PlexConnectATVLanguage'] = language
-                            break
                     XML = XMLConverter.XML_PMS2aTV(self.client_address, self.path, options)
                     self.send_response(200)
                     self.send_header('Content-type', 'text/xml')
