+++ assets/templates/Season_List.xml
@@ -17,7 +17,7 @@
             <items>
               <twoLineEnhancedMenuItem id="{{VAL(key)}}" onPlay="atv.loadURL('{{URL(key)}}')"
                                                          onSelect="atv.loadURL('{{URL(key)}}')"
-                                                         onHoldSelect="scrobbleMenu('Season', '{{VAL(ratingKey)}}', '{{ADDR_PMS()}}');">
+                                                         onHoldSelect="scrobbleMenu('{{TEXT(Season)}}', '{{VAL(ratingKey)}}', '{{ADDR_PMS()}}');">
                 {{COPY(Directory)}}
                 <label>{{VAL(title)}}</label>
                 <label2>{{VAL(leafCount)}} {{TEXT(Episodes)}}</label2>
