+++ assets/templates/Show_Grid.xml
@@ -18,7 +18,7 @@
             <moviePoster id="{{VAL(key)}}"
                          onPlay="atv.loadURL('{{URL(key)}}')"
                          onSelect="atv.loadURL('{{URL(key)}}')"
-                         onHoldSelect="scrobbleMenu('Show', '{{VAL(ratingKey)}}', '{{ADDR_PMS()}}');">
+                         onHoldSelect="scrobbleMenu('{{TEXT(Show)}}', '{{VAL(ratingKey)}}', '{{ADDR_PMS()}}');">
               {{COPY(Directory)}}
               <title>{{VAL(title)}}</title>
               <image>{{IMAGEURL(thumb)}}</image>
